refactor(schemas): migrate Phase 23 endpoint schemas — 5 files, 15 classes (#6042)
- a2a.py (3) → schemas_agent.py
- agents_self_improvement.py (3) → schemas_agent.py
- enhanced_memory.py (3) → schemas_knowledge.py
- integration_cicd.py (3) → schemas_workflows.py (renames: ConnectionTestRequest
  → CICDConnectionTestRequest, ProviderInfo → CICDProviderInfo)
- enterprise_features.py (3) → schemas_workflows.py

Renames to avoid generic-name collisions:
- integration_cicd's ConnectionTestRequest → CICDConnectionTestRequest
- integration_cicd's ProviderInfo → CICDProviderInfo

CICDProvider Literal type moved to schemas_workflows.py for centralized use.
